requirements.txt
SQLAlchemy==1.4.28
google-cloud==0.34.0
google-cloud-pubsub==2.9.0
Flask==2.0.2
functions-framework==3.*
pg8000==1.23.0